[ZenCPU] Add separate ZentorchWNA16LinearKernel
- Add new Zentorch W4A16 weight-only-quantized linear kernel
- Kernel is selected by `choose_mp_linear_kernel` ahead of the
  generic oneDNN-backed `CPUWNA16LinearKernel`. When `can_implement`
  reject a layer, the selector falls through to the next kernel in
  `_POSSIBLE_KERNELS[PlatformEnum.CPU]`.
